Re-organize the bigger value package
Added some comments to undocumented methods, and also split the now too
big value.go file into smaller files.
